CUDA: Fix various issues in test suite:
- Replace deprecated `argtypes` / `restypes` use with signatures:
  - `test_blackscholes.py`
  - `test_idiv.py`
  - `test_matmul.py`
  - `test_nondet.py`
  - `test_powi.py`
  - `test_py2_div_issue.py`
- Correct mismatches between signatures and arguments:
  - `test_linker.py`
  - `test_constmem.py`
  - `test_intrinsics.py`
- Correct logic errors:
  - `test_optimization.py`: This test didn't correctly get the PTX, and
     was instead asserting that the second part of the key from
     `inspect_asm()` (the signature) didn't contain the removed
     fragments, instead of actually checking the PTX code.

+++ numba/cuda/tests/cudadrv/test_linker.py
@@ -78,8 +78,8 @@ class TestLinker(CUDATestCase):
             i = cuda.grid(1)
             x[i] += bar(y[i])
 
-        A = np.array([123])
-        B = np.array([321])
+        A = np.array([123]).astype(np.int32)
+        B = np.array([321]).astype(np.int32)
 
         foo[1, 1](A, B)
